+++ test/algorithms/refinement/tst_orientation_refinement.py
@@ -194,7 +194,7 @@ obs_refs = obs_refs.select(intersects)
 # Make a reflection predictor and re-predict for all these reflections. The
 # result is the same, but we gain also the flags and xyzcal.px columns
 ref_predictor = ExperimentsPredictor(experiments)
-obs_refs['id'] = flex.size_t(len(obs_refs), 0)
+obs_refs['id'] = flex.int(len(obs_refs), 0)
 obs_refs = ref_predictor.predict(obs_refs)
 
 # Set 'observed' centroids from the predicted ones
